# taskify-front/assets/openapi.json
{
    "openapi": "3.1.0",
    "info": {
    "title": "Taskify Backend",
    "version": "0.1.0"
    },
    "paths": {
...
    "/api/v1/users/": {
    "get": {
    "tags": [
    "Users"
    ],
    "summary": "Usercontroller.Filter Users",
    "operationId": "UserController_filter_users_api_v1_users__get",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "email",
    "in": "query",
    "required": false,
    "schema": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Email"
    }
    },
    {
    "name": "page",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "minimum": 1,
    "default": 1,
    "title": "Page"
    }
    },
    {
    "name": "size",
    "in": "query",
    "required": false,
    "schema": {
    "type": "integer",
    "maximum": 100,
    "minimum": 1,
    "default": 50,
    "title": "Size"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/Page_BoardMemeber_"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/users/me": {
    "get": {
    "tags": [
    "Users"
    ],
    "summary": "Usercontroller.Get Me",
    "operationId": "UserController_get_me_api_v1_users_me_get",
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/UserRead"
    }
    }
    }
    }
    },
    "security": [
    {
    "APIKeyCookie": []
    }
    ]
    },
    "patch": {
    "tags": [
    "Users"
    ],
    "summary": "Usercontroller.Patch Me",
    "operationId": "UserController_patch_me_api_v1_users_me_patch",
    "requestBody": {
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/UserUpdate"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/UserRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    },
    "security": [
    {
    "APIKeyCookie": []
    }
    ]
    }
    },
    "/api/v1/users/{user_id}": {
    "get": {
    "tags": [
    "Users"
    ],
    "summary": "Usercontroller.Get User",
    "operationId": "UserController_get_user_api_v1_users__user_id__get",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "user_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "User Id"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/UserRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
...
    "/api/v1/oauth/github/callback": {
    "get": {
    "tags": [
    "Auth",
    "OAuth"
    ],
    "summary": "Github-Callback",
    "operationId": "github_callback_api_v1_oauth_github_callback_get",
    "parameters": [
    {
    "name": "code",
    "in": "query",
    "required": false,
    "schema": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Code"
    }
    },
    {
    "name": "code_verifier",
    "in": "query",
    "required": false,
    "schema": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Code Verifier"
    }
    },
    {
    "name": "state",
    "in": "query",
    "required": false,
    "schema": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "State"
    }
    },
    {
    "name": "error",
    "in": "query",
    "required": false,
    "schema": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Error"
    }
    }
    ],
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {}
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
...
    "/api/v1/oauth/{oauth_id}": {
    "delete": {
    "tags": [
    "Auth",
    "OAuth"
    ],
    "summary": "Oauthcontroller.Remove Oauth Account",
    "operationId": "OAuthController_remove_oauth_account_api_v1_oauth__oauth_id__delete",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "oauth_id",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "format": "uuid",
    "title": "Oauth Id"
    }
    }
    ],
    "responses": {
    "204": {
    "description": "Successful Response"
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    },
    "/api/v1/files/avatars": {
    "put": {
    "tags": [
    "Files"
    ],
    "summary": "Filestoragecontroller.Upload Avatar",
    "operationId": "FileStorageController_upload_avatar_api_v1_files_avatars_put",
    "requestBody": {
    "content": {
    "multipart/form-data": {
    "schema": {
    "$ref": "#/components/schemas/Body_FileStorageController_upload_avatar_api_v1_files_avatars_put"
    }
    }
    },
    "required": true
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/UserRead"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    },
    "security": [
    {
    "APIKeyCookie": []
    }
    ]
    }
    },
    "/api/v1/files/{object_key}": {
    "put": {
    "tags": [
    "Files"
    ],
    "summary": "Filestoragecontroller.Upload File",
    "operationId": "FileStorageController_upload_file_api_v1_files__object_key__put",
    "security": [
    {
    "APIKeyCookie": []
    }
    ],
    "parameters": [
    {
    "name": "object_key",
    "in": "path",
    "required": true,
    "schema": {
    "type": "string",
    "title": "Object Key"
    }
    }
    ],
    "requestBody": {
    "required": true,
    "content": {
    "multipart/form-data": {
    "schema": {
    "$ref": "#/components/schemas/Body_FileStorageController_upload_file_api_v1_files__object_key__put"
    }
    }
    }
    },
    "responses": {
    "200": {
    "description": "Successful Response",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/FileUploadResponse"
    }
    }
    }
    },
    "422": {
    "description": "Validation Error",
    "content": {
    "application/json": {
    "schema": {
    "$ref": "#/components/schemas/HTTPValidationError"
    }
    }
    }
    }
    }
    }
    }
    },
    "components": {
    "schemas": {
...
    "BoardMemeber": {
    "properties": {
    "id": {
    "type": "string",
    "format": "uuid",
    "title": "Id"
    },
    "created_at": {
    "type": "string",
    "format": "date-time",
    "title": "Created At"
    },
    "updated_at": {
    "type": "string",
    "format": "date-time",
    "title": "Updated At"
    },
    "email": {
    "type": "string",
    "title": "Email"
    },
    "username": {
    "type": "string",
    "title": "Username"
    },
    "full_name": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Full Name"
    },
    "avatar": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Avatar"
    }
    },
    "type": "object",
    "required": [
    "id",
    "created_at",
    "updated_at",
    "email",
    "username",
    "full_name",
    "avatar"
    ],
    "title": "BoardMemeber"
    },
...
    "Body_FileStorageController_upload_avatar_api_v1_files_avatars_put": {
    "properties": {
    "file": {
    "type": "string",
    "format": "binary",
    "title": "File"
    }
    },
    "type": "object",
    "required": [
    "file"
    ],
    "title": "Body_FileStorageController_upload_avatar_api_v1_files_avatars_put"
    },
    "Body_FileStorageController_upload_file_api_v1_files__object_key__put": {
    "properties": {
    "file": {
    "type": "string",
    "format": "binary",
    "title": "File"
    }
    },
    "type": "object",
    "required": [
    "file"
    ],
    "title": "Body_FileStorageController_upload_file_api_v1_files__object_key__put"
    },
...
    "FileUploadResponse": {
    "properties": {
    "filename": {
    "type": "string",
    "title": "Filename"
    },
    "bucket": {
    "type": "string",
    "title": "Bucket",
    "default": "taskifybucket"
    },
    "object_key": {
    "type": "string",
    "title": "Object Key"
    },
    "domain_url": {
    "type": "string",
    "title": "Domain Url"
    },
    "size": {
    "type": "integer",
    "title": "Size",
    "default": 0
    },
    "url": {
    "type": "string",
    "title": "Url"
    }
    },
    "type": "object",
    "required": [
    "filename",
    "object_key",
    "domain_url"
    ],
    "title": "FileUploadResponse"
    },
...
    "HTTPValidationError": {
    "properties": {
    "detail": {
    "items": {
    "$ref": "#/components/schemas/ValidationError"
    },
    "type": "array",
    "title": "Detail"
    }
    },
    "type": "object",
    "title": "HTTPValidationError"
    },
...
    "OAuthRead": {
    "properties": {
    "id": {
    "type": "string",
    "format": "uuid",
    "title": "Id"
    },
    "user_id": {
    "type": "string",
    "format": "uuid",
    "title": "User Id"
    },
    "oauth_name": {
    "type": "string",
    "title": "Oauth Name"
    },
    "account_id": {
    "type": "string",
    "title": "Account Id"
    },
    "account_email": {
    "type": "string",
    "title": "Account Email"
    }
    },
    "type": "object",
    "required": [
    "id",
    "user_id",
    "oauth_name",
    "account_id",
    "account_email"
    ],
    "title": "OAuthRead"
    },
    "Page_BoardMemeber_": {
    "properties": {
    "items": {
    "items": {
    "$ref": "#/components/schemas/BoardMemeber"
    },
    "type": "array",
    "title": "Items"
    },
    "total": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Total"
    },
    "page": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Page"
    },
    "size": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 1
    },
    {
    "type": "null"
    }
    ],
    "title": "Size"
    },
    "pages": {
    "anyOf": [
    {
    "type": "integer",
    "minimum": 0
    },
    {
    "type": "null"
    }
    ],
    "title": "Pages"
    }
    },
    "type": "object",
    "required": [
    "items",
    "page",
    "size"
    ],
    "title": "Page[BoardMemeber]"
    },
...
    "UserRead": {
    "properties": {
    "id": {
    "type": "string",
    "format": "uuid",
    "title": "Id"
    },
    "created_at": {
    "type": "string",
    "format": "date-time",
    "title": "Created At"
    },
    "updated_at": {
    "type": "string",
    "format": "date-time",
    "title": "Updated At"
    },
    "email": {
    "type": "string",
    "title": "Email"
    },
    "username": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Username"
    },
    "full_name": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Full Name"
    },
    "avatar": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Avatar"
    },
    "role_id": {
    "type": "integer",
    "title": "Role Id"
    },
    "is_active": {
    "type": "boolean",
    "title": "Is Active"
    },
    "oauth_accounts": {
    "items": {
    "$ref": "#/components/schemas/OAuthRead"
    },
    "type": "array",
    "title": "Oauth Accounts",
    "default": []
    }
    },
    "type": "object",
    "required": [
    "id",
    "created_at",
    "updated_at",
    "email",
    "username",
    "full_name",
    "avatar",
    "role_id",
    "is_active"
    ],
    "title": "UserRead"
    },
    "UserUpdate": {
    "properties": {
    "email": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Email"
    },
    "username": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Username"
    },
    "full_name": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "null"
    }
    ],
    "title": "Full Name"
    }
    },
    "type": "object",
    "title": "UserUpdate"
    },
    "ValidationError": {
    "properties": {
    "loc": {
    "items": {
    "anyOf": [
    {
    "type": "string"
    },
    {
    "type": "integer"
    }
    ]
    },
    "type": "array",
    "title": "Location"
    },
    "msg": {
    "type": "string",
    "title": "Message"
    },
    "type": {
    "type": "string",
    "title": "Error Type"
    }
    },
    "type": "object",
    "required": [
    "loc",
    "msg",
    "type"
    ],
    "title": "ValidationError"
    }
    },
    "securitySchemes": {
    "APIKeyCookie": {
    "type": "apiKey",
    "in": "cookie",
    "name": "taskify_auth"
    }
    }
    }
    }